Change API Id for group and clean index.js and test new Id to be working when adding reservation
