+++ src/modules/getReserve.js
@@ -1,21 +1,20 @@
 import * as resInterface from "./reserveInterface.js";
 
-const projectId = 'NXpRpTwwlkHX1mjqefPA';
-const url ='https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/NXpRpTwwlkHX1mjqefPA/reservations?item_id=item1/';
+const projectId = 'HrIKPRrYjrxS00NlIVCD';
+const url ='https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/HrIKPRrYjrxS00NlIVCD/reservations?item_id=item1/';
 
-const getReserveData = async () => {
-  const showList = resInterface.resListContent;
-    let resultData;
-  try {
-  const response = await fetch(url);
-  const result = await response.json();
-    resultData = result;
-      showList.innerText = "No reservations";
-  } catch (error) {
-    console.log(error.message);
-  }
- // console.log(resultData);
- return resultData;
+const getReserveData = async (imgId) => {
+console.log(imgId)
+//     let resultData;
+//   try {
+//   const response = await fetch(url);
+//   const result = await response.json();
+//     resultData = result;
+//       showList.innerText = "No reservations";
+//   } catch (error) {
+//     console.log(error.message);
+//   }
+//  return resultData;
 }
 
 export default getReserveData;
