--- src/modules/popupElements.js
@@ -1,10 +1,0 @@
-
-<img src="${theimgs[btnIndex]}" class="dog-image" alt="image of the selected dog"/>
-<h2 class="reserve-data">Reservations (data)</h2>
-<div class="reserve-list">use paragraph here</div>
-<form id="reserve-form" class="reserve-form">
-    <input type="text" id="username" class="username" value="" maxlength="40" placeholder="Your name">
-    <input type="date" id="startdate" class="startdate" value="" placeholder="Start date">
-    <input type="date" id="enddate" class="enddate" value="" placeholder="End date">
-    <button type="button" id="submit" class="submit" value="submit">Reserve</button>
-</form>
